Auth (#3)
* Add users route and auth for users

* Use passport isAuthenticated helper

Require users to be signed in in order to create a new campground

* Move isAuthenticated into middleware

Remove passport helper isAuthenicated from campgrounds.js into a middleware that can be reused.

* Protect campground routes

Add middleware to ensure users are logged in in order to add/edit/delete campgrounds.

* Auth links

Add links for login/logout/register. Add route to logout.

* Auth links visibility

Add the currentUser to the locals array in app.js and show/hide auth links based on whether the user is logged in or not.

* Fix register route

Log the newly registered user in so they don't have to log in again to be fully logged in

* returnTo url

Add a returnTo property to the session in order to know where the user was originally trying to go when logging in

* Fix login bug

fix login bug where you could still be redirected to /campgrounds/new after logging in even if you weren't logging in from this route.